genesis-poc/requirements.txt
aws-opentelemetry-distro
pydantic
boto3
botocore
langchain-aws
langchain-community
langgraph
asciitree
flask[async]
flask-socketio
aiofiles
google-search-results
langchain-mcp-adapters
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
click==8.1.8
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
mcp==1.4.1
pydantic-settings==2.8.1
pydantic_core==2.27.2
python-dotenv==1.0.1
sniffio==1.3.1
sse-starlette==2.2.1
starlette==0.46.1
typing_extensions==4.12.2
uvicorn==0.34.0
langgraph_checkpoint_aws
opentelemetry-api
opentelemetry-sdk
pytz
pandas
opentelemetry-instrumentation-langchain
langchain-openai
opentelemetry-exporter-otlp
fastapi
